+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -230,7 +230,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K5V5RXZ7FXMDWX80CJBS7RQY"
+        "01K5X16BZMYXH42ZZACKKE4YPB"
     )
     r"""The id of the resource"""