+++ packages/orq-rc/src/orq_ai_sdk/models/listcontactsop.py
@@ -184,7 +184,7 @@ class Data(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-09-23T11:36:26.030Z")
+    updated: Optional[datetime] = parse_datetime("2025-09-24T04:54:52.401Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")