+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -132,7 +132,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K5V5RXZ5GFTGF3EGHYRAEQ1A"
+        "01K5X16BZJ8KGHD40WN8B0K0TS"
     )
     r"""The id of the resource"""