+++ packages/orq-rc/RELEASES.md
@@ -2879,3 +2879,13 @@ Based on:
 - [python v3.13.0-rc.29] packages/orq-rc
 ### Releases
 - [PyPI v3.13.0-rc.29] https://pypi.org/project/orq-ai-sdk/3.13.0-rc.29 - packages/orq-rc
+
+## 2025-09-24 04:56:33
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.624.0 (2.714.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.13.0-rc.30] packages/orq-rc
+### Releases
+- [PyPI v3.13.0-rc.30] https://pypi.org/project/orq-ai-sdk/3.13.0-rc.30 - packages/orq-rc